provisioning_service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
asyncpg>=0.29.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
celery>=5.3.0
pydantic[email]>=2.0.0
sqlalchemy>=2.0.0
resend>=2.0.0
